requirements.txt
requests==2.31.0
python-krb5ticket==0.1.2
